+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+- Animations of moving from one field to another
+- Persons with same weapons rendered on stage when attack
+
 ## [1.1] - 2022-03-12
 - Fixed scissors not cutting paper -> paper won in that case